\name{ddvCopula-methods}
\docType{methods}
\alias{ddvCopula-methods}

\alias{ddvCopula,matrix,claytonCopula-method}
\alias{ddvCopula,matrix,frankCopula-method}
\alias{ddvCopula,matrix,gumbelCopula-method}
\alias{ddvCopula,matrix,indepCopula-method}
\alias{ddvCopula,matrix,normalCopula-method}
\alias{ddvCopula,matrix,tCopula-method}
\alias{ddvCopula,matrix,leafCopula-method}

\alias{ddvCopula,numeric,claytonCopula-method}
\alias{ddvCopula,numeric,frankCopula-method}
\alias{ddvCopula,numeric,gumbelCopula-method}
\alias{ddvCopula,numeric,indepCopula-method}
\alias{ddvCopula,numeric,normalCopula-method}
\alias{ddvCopula,numeric,tCopula-method}
\alias{ddvCopula,numeric,leafCopula-method}

\title{Methods for Function \code{ddvCopula} in Package \pkg{sfcopula}}
\description{
 Methods for function \code{ddvCopula} in package \pkg{sfcopula}
}
\section{Methods}{
\describe{

\item{\code{signature(u = "matrix", copula = "claytonCopula")}}{
}
\item{\code{signature(u = "matrix", copula = "frankCopula")}}{
}
\item{\code{signature(u = "matrix", copula = "gumbelCopula")}}{
}
\item{\code{signature(u = "matrix", copula = "indepCopula")}}{
}
\item{\code{signature(u = "matrix", copula = "normalCopula")}}{
}
\item{\code{signature(u = "matrix", copula = "tCopula")}}{
}

\item{\code{signature(u = "numeric", copula = "claytonCopula")}}{
}
\item{\code{signature(u = "numeric", copula = "frankCopula")}}{
}
\item{\code{signature(u = "numeric", copula = "gumbelCopula")}}{
}
\item{\code{signature(u = "numeric", copula = "indepCopula")}}{
}
\item{\code{signature(u = "numeric", copula = "normalCopula")}}{
}
\item{\code{signature(u = "numeric", copula = "tCopula")}}{
}
}}
\keyword{methods}
\keyword{partial derivatives}
